lib: Provide cockpit.js as ES6 module
Move cockpit.js' code to pkg/lib and provide its object as default
export. With that, external projects like d-installer [1] can import it
through webpack/ES6 instead of `<script>`.

Only keep the cockpit.css build in base1/cockpit.js, as that's not
relevant for using it as a module.

It would also be elegant to keep the `window.cockpit` global definition
in base1/cockpit.js, but that would break translations: Importing
`po.js` through `<script>` must have access to the `cockpit` global
already, and po.js *has* to be loaded that way due to the dynamic locale
specific path lookup.

Fixes #17203

[1] https://github.com/yast/d-installer
