+++ containers/unit-tests/scenario/distcheck
@@ -6,7 +6,7 @@
 # check translation build
 make po/cockpit.pot
 # do some spot checks
-grep -q 'pkg/base1/cockpit.js' po/cockpit.pot
+grep -q 'pkg/lib/cockpit.js' po/cockpit.pot
 grep -q 'pkg/shell/hosts_dialog.jsx' po/cockpit.pot
 grep -q 'pkg/systemd/services.html' po/cockpit.pot
 grep -q 'pkg/static/login.html' po/cockpit.pot
